+++ src/ShellCheck/Parser.hs
@@ -2088,7 +2088,8 @@ readSource t@(T_Redirecting _ _ (T_SimpleCommand cmdId _ (cmd:file':rest'))) = d
                     if filename == "/dev/null" -- always allow /dev/null
                     then return (Right "")
                     else do
-                        filename' <- system $ siFindSource sys filename
+                        currentScript <- Mr.asks currentFilename
+                        filename' <- system $ siFindSource sys currentScript filename
                         system $ siReadFile sys filename'
                 case input of
                     Left err -> do